--- a/colcon_python_setup_py/package_identification/run_setup_py.py
+++ b/colcon_python_setup_py/package_identification/run_setup_py.py
@@ -1,3 +1,6 @@
+# Copyright 2019 Rover Robotics via Dan Rose
+# Licensed under the Apache License, Version 2.0
+
 import distutils.core
 import os