1. Move dependency on setuptools closer to where it's needed
2. Scope multiprocessing Pool so a pool left open doesn't hang the process
3. Move run_setup_py to its own file so the subprocess doesn't need to import much

I think it would be nice for colcon to pass in a ProcessPoolExecutor that has a longer life, so we don't need to spin one up so often (the expensive part).