@dirk-thomas 2019-10-23
This (together with colcon/colcon-ros#86) fails to build a ROS 2 workspace for me:

```
Traceback (most recent call last):
  File "/home/dthomas/colcon/ws/build/colcon-core/colcon_core/executor/__init__.py", line 91, in __call__
    rc = await self.task(*args, **kwargs)
  File "/home/dthomas/colcon/ws/build/colcon-core/colcon_core/task/__init__.py", line 92, in __call__
    return await task_method(*args, **kwargs)
  File "/home/dthomas/colcon/ws/build/colcon-ros/colcon_ros/task/ament_python/build.py", line 97, in build
    return await extension.build(additional_hooks=additional_hooks)
  File "/home/dthomas/colcon/ws/build/colcon-core/colcon_core/task/python/build.py", line 82, in build
    self._symlinks_in_build(args, setup_py_data)
  File "/home/dthomas/colcon/ws/build/colcon-core/colcon_core/task/python/build.py", line 188, in _symlinks_in_build
    if package in package_dir:
TypeError: argument of type 'NoneType' is not iterable
```