@dirk-thomas 2019-10-24
How does this always re-instantiated multiprocessing pool perform compared to a plain subprocess invocation in terms of runtime?