@rotu 2019-10-24
Significant, but not crazy. It's 4x slower than spawning a Python interpreter, but if you reuse the process pool, it's 7x faster.

> call 1.0819000000009127e-07
> subprocess 0.00517880357
> interpreter 0.0331634016
> transient pool 0.11082568893
> shared pool 0.004250488220000008

```python3
import subprocess
import multiprocessing
import gc
from timeit import timeit

def f0(): return True
def f1(): subprocess.call(['true'])
def f2(): subprocess.call(['python3', '-c', 'True'])
def f3():
    with multiprocessing.Pool(1) as pool: pool.apply(f1)
def f4(p): p.apply(f1)

setup = 'gc.enable()'
n=100
print('call', timeit('f0()', setup, number=n, globals=globals())/n)
print('subprocess', timeit('f1()', setup, number=n, globals=globals())/n)
print('interpreter', timeit('f2()', setup, number=n, globals=globals())/n)
print('transient pool', timeit('f3()', setup, number=n, globals=globals())/n)
with multiprocessing.Pool(1) as pool:
    print('shared pool', timeit('f4(pool)', setup, number=n,
        globals=globals())/n)
```