@dirk-thomas 2019-10-24
> It's 4x slower than spawning a Python interpreter in a new process, but if you reuse the process pool, it's 7x faster.

Since the patch currently doesn't reuse the process pool how about keep using subprocess as it is right now in the target branch?