@rotu 2019-10-24
What like about the process pool is it handles return values and exceptions. Plus passing in a process pool can speed up this code as well as the Python project build code.

I think it’s worth it to keep using ProcessPool, even if it is a bit slower for now. Premature optimization and whatnot.