@dirk-thomas 2019-10-24
> Plus passing in a process pool can speed up this code as well as the Python project build code.

Can you describe how you think the process pool can be used in the future?

> Premature optimization and whatnot.

This isn't premature optimization. The target branch uses subprocess and the proposed change switches it to a one-time process pool which is a regression in performance (I might need to measure how much it changes). So at the moment using the process pool provides only some convenience advantage which isn't a good argument for the performance impact (assuming it is not insignificant).