@rotu 2019-10-24
> Can you describe how you think the process pool can be used in the future?

Sure. Passing in a process pool to PythonBuildTask and PythonTestTask prevents the need to spin up a process and (more significantly) the Python interpreter for each invocation. This could be created in the context and persisted for the entire run of colcon.

> This isn't premature optimization.

How noticeable is the change in performance? I actually observe a *speedup* with `colcon list` with the unoptimised new version:

before:
colcon list  2.74s user 0.57s system 86% cpu 3.822 total

after:
colcon list  2.52s user 0.42s system 90% cpu 3.257 total