@dirk-thomas 2019-10-24
I compared the timing of `colcon list` for a larger ROS 2 workspace contain several Python packages and this approach seems to take pretty much the same time as the existing one. So I am ok to merge this when you are.