@rotu 2019-10-25
After one last fix (see prev commit), I'm ready to merge this.
If performance is a problem, #25 will make it faster, though still not as fast as a process pool.