@dirk-thomas 2020-04-01
The target branch of this PR only existed while we worked on different approaches to get the information from Python setup files. In the meantime #30 has been merged so I don't see how the current patch can be applied anymore.

Therefore I will go ahead and close this ticket. If necessary please consider creating a new pull request targeting `master` for any left over desired changes.